+ restructure and working of different registries
