+++ app/apis.registry.js
@@ -1,0 +1,1 @@
+export const ApiRegistry = {};
