+++ app/formValidationMap.registry.js
@@ -1,0 +1,1 @@
+export const ValidationRegistry = {};
