+++ app/functions/asyncSelect.functions.js
@@ -1,0 +1,11 @@
+export const masterData = {
+  getOptionLabel: (data) => {
+    return data?.label || "";
+  },
+  getOptionValue: (data) => {
+    return data?.name || "";
+  },
+  isOptionsEqualToValue: (option, value) => {
+    return option?.name === value;
+  },
+};
